Released v1.00a: Minor details fixed
Thanks to Ulrich Diez for explaining this LaTeX Warning (by pdflatex counttexruns.dtx):
| LaTeX Warning: You have requested package `counttexruns',
|                 but the package provides `counttexruns'.

The reason is:

In the dtx-file you find (line 41):  \usepackage{\jobname}  .
In the sty-file you find (line 3):  \ProvidesPackage{counttexruns}  .

\jobname yields a set of catcode-12-character-tokens and space-tokens.

Tokenizing the phrase "counttexruns" when reading the sty-file yields
a set of catcode-11-character-tokens.

Thus the two sets of tokens differ although they look the same
when written on the screen.

Either within the dtx-file define a macro
\ThisPackagesLovelyName to expand to the phrase "counttexruns"
and use that instead of \jobname.

Or use \scantokens for retokenization of \jobname's expansion
when calling the package in line 41 of the dtx-file:

\scantokens\expandafter{\expandafter\usepackage\expandafter{\jobname}}%