Fix compatibility issue with newer versions of hyperxmp. Closes #66.
Load order must be precise because hyperxmp also installs a hook with \AtEndPreamble same as we do. Now our hook needs to be executed before the one of hyperxmp because hyperxmp checks if hyperref is loaded and if it does not at the point where hyperxmp checks, it will do nothing.