chore(doclicense): Do not try to load hyperxmp if silence is loaded
I am not keen to fix this in doclicense. But the facts that it worked
before (because we did not load hyperxmp by default) and the fact that
this is some sort of endless loop that could otherwise be hard to debug
and waste CPU time and therefore resources, I decided to catch this
case.

Closes: #78